+++ docs/changelog.rst
@@ -47,6 +47,8 @@ Detailed list of changes
 
 - kitten @ ls: Allow limiting output to matched windows/tabs (:iss:`6520`)
 
+- Remote control: Allow using a random TCP socket as the remote control socket and also allow using TCP sockets in :opt:`listen_on`
+
 - X11: Print an error to :file:`STDERR` instead of refusing to start when the user sets a custom window icon larger than 128x128 (:iss:`6507`)
 
 0.29.2 [2023-07-27]